# backend/openapi-live.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "BookSpot API",
    "description": "A comprehensive booking system API for service providers and clients",
    "contact": {
      "name": "BookSpot Support",
      "email": "support@bookspot.com"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "v1"
  },
  "paths": {
...
    "/dashboard/me": {
      "get": {
        "tags": [
          "Dashboard"
        ],
        "summary": "GetMine Dashboard",
        "description": "Performs GetMine operation on Dashboard",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ClientDashboardDto"
                    },
                    {
                      "$ref": "#/components/schemas/ProviderDashboardDto"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing JWT token",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Insufficient permissions",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Validation errors or invalid input",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/dashboard/provider/{providerId}/stats": {
      "get": {
        "tags": [
          "Dashboard"
        ],
        "summary": "GetProviderStats Dashboard",
        "description": "Performs GetProviderStats operation on Dashboard",
        "parameters": [
          {
            "name": "providerId",
            "in": "path",
            "description": "Provider's user ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Statistics retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardStats"
                }
              }
            }
          },
          "404": {
            "description": "Provider not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "Invalid provider ID",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - JWT token required",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Only providers can access dashboard stats",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/dashboard/providers/{providerId}/insights": {
      "get": {
        "tags": [
          "Dashboard"
        ],
        "summary": "GetProviderInsights Dashboard",
        "description": "Performs GetProviderInsights operation on Dashboard",
        "parameters": [
          {
            "name": "providerId",
            "in": "path",
            "description": "Provider's user ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "in": "query",
            "description": "Filter bookings from this date (inclusive)",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "description": "Filter bookings up to this date (inclusive)",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Insights retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderInsightsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid filter values",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated user does not match providerId",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Provider not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/dashboard/clients": {
      "get": {
        "tags": [
          "Dashboard"
        ],
        "summary": "GetClients Dashboard",
        "description": "Performs GetClients operation on Dashboard",
        "responses": {
          "200": {
            "description": "Clients retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DashboardClientDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Provider is not recognized as a service provider",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Only providers can access client records",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Provider profile not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/dashboard/client/{clientId}/stats": {
      "get": {
        "tags": [
          "Dashboard"
        ],
        "summary": "GetClientStats Dashboard",
        "description": "Performs GetClientStats operation on Dashboard",
        "parameters": [
          {
            "name": "clientId",
            "in": "path",
            "description": "Client's user ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Client statistics retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientStatsDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid client ID or user is not a client",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Access denied - clients can only view their own stats",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Client not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/dashboard/my-stats": {
      "get": {
        "tags": [
          "Dashboard"
        ],
        "summary": "GetMyStats Dashboard",
        "description": "Performs GetMyStats operation on Dashboard",
        "responses": {
          "200": {
            "description": "Personal statistics retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientStatsDto"
                }
              }
            }
          },
          "400": {
            "description": "User is not a client",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Only clients can access their own statistics",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Client profile not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ClientDashboardDto": {
        "type": "object",
        "allOf": [
          {
            "$ref": "#/components/schemas/DashboardDto"
          }
        ],
        "properties": {
          "generatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "totalBookings": {
            "type": "integer",
            "format": "int32"
          },
          "completedBookings": {
            "type": "integer",
            "format": "int32"
          },
          "cancelledBookings": {
            "type": "integer",
            "format": "int32"
          },
          "pendingRequests": {
            "type": "integer",
            "format": "int32"
          },
          "totalSpent": {
            "$ref": "#/components/schemas/MoneyDto"
          },
          "upcoming": {
            "type": "array",
            "items": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/ClientBookingDto"
                },
                {
                  "$ref": "#/components/schemas/ProviderBookingDto"
                }
              ]
            },
            "nullable": true
          },
          "recent": {
            "type": "array",
            "items": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/ClientBookingDto"
                },
                {
                  "$ref": "#/components/schemas/ProviderBookingDto"
                }
              ]
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ClientStatsDto": {
        "type": "object",
        "properties": {
          "clientId": {
            "type": "string",
            "nullable": true
          },
          "fullName": {
            "type": "string",
            "nullable": true,
            "example": "John Doe"
          },
          "email": {
            "type": "string",
            "nullable": true,
            "example": "user@example.com"
          },
          "contactNumber": {
            "type": "string",
            "nullable": true,
            "example": "+1 (555) 123-4567"
          },
          "totalBookings": {
            "type": "integer",
            "format": "int32"
          },
          "completedBookings": {
            "type": "integer",
            "format": "int32"
          },
          "cancelledBookings": {
            "type": "integer",
            "format": "int32"
          },
          "totalSpent": {
            "type": "number",
            "format": "double"
          },
          "firstVisit": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "lastVisit": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "favoriteService": {
            "type": "string",
            "nullable": true
          },
          "recentBookings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RecentBookingDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "DashboardClientDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "fullName": {
            "type": "string",
            "nullable": true,
            "example": "John Doe"
          },
          "email": {
            "type": "string",
            "nullable": true,
            "example": "user@example.com"
          },
          "contactNumber": {
            "type": "string",
            "nullable": true,
            "example": "+1 (555) 123-4567"
          },
          "totalBookings": {
            "type": "integer",
            "format": "int32"
          },
          "lastVisit": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "DashboardStats": {
        "type": "object",
        "properties": {
          "todayBookings": {
            "type": "integer",
            "format": "int32"
          },
          "weekBookings": {
            "type": "integer",
            "format": "int32"
          },
          "totalClients": {
            "type": "integer",
            "format": "int32"
          },
          "monthlyRevenue": {
            "type": "number",
            "format": "double"
          },
          "pendingBookings": {
            "type": "integer",
            "format": "int32"
          },
          "confirmedBookings": {
            "type": "integer",
            "format": "int32"
          },
          "completedBookings": {
            "type": "integer",
            "format": "int32"
          },
          "cancelledBookings": {
            "type": "integer",
            "format": "int32"
          },
          "averageBookingValue": {
            "type": "number",
            "format": "double"
          },
          "totalBookings": {
            "type": "integer",
            "format": "int32"
          },
          "statsGeneratedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
      "ProviderDashboardDto": {
        "type": "object",
        "allOf": [
          {
            "$ref": "#/components/schemas/DashboardDto"
          }
        ],
        "properties": {
          "generatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "timeZone": {
            "type": "string",
            "nullable": true
          },
          "todayBookings": {
            "type": "integer",
            "format": "int32"
          },
          "weekBookings": {
            "type": "integer",
            "format": "int32"
          },
          "pendingRequests": {
            "type": "integer",
            "format": "int32"
          },
          "totalClients": {
            "type": "integer",
            "format": "int32"
          },
          "activeServices": {
            "type": "integer",
            "format": "int32"
          },
          "monthlyRevenue": {
            "$ref": "#/components/schemas/MoneyDto"
          },
          "upcoming": {
            "type": "array",
            "items": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/ClientBookingDto"
                },
                {
                  "$ref": "#/components/schemas/ProviderBookingDto"
                }
              ]
            },
            "nullable": true
          },
          "recentClients": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RecentClientDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ProviderInsightsResponse": {
        "type": "object",
        "properties": {
          "stats": {
            "$ref": "#/components/schemas/ProviderInsightsStats"
          },
          "popularServices": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PopularServiceInsight"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "JWT Authorization header using the Bearer scheme. Enter 'Bearer' [space] and then your token in the text input below.",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}